Move code for the GradeBook GUI to its own package.
