+++ grader/src/main/java/edu/pdx/cs410J/grader/GraderTools.java
@@ -4,6 +4,7 @@ import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import com.google.common.annotations.VisibleForTesting;
 import edu.pdx.cs410J.grader.canvas.GradesFromCanvasImporter;
+import edu.pdx.cs410J.grader.gradebook.ui.GradeBookGUI;
 import edu.pdx.cs410J.grader.poa.ui.PlanOfAttackGrader;
 import org.slf4j.LoggerFactory;
