+++ grader/src/main/java/edu/pdx/cs410J/grader/gradebook/ui/GradeBookGUI.java
@@ -1,11 +1,12 @@
-package edu.pdx.cs410J.grader;
+package edu.pdx.cs410J.grader.gradebook.ui;
 
-import java.awt.BorderLayout;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.awt.event.KeyEvent;
-import java.awt.event.WindowAdapter;
-import java.awt.event.WindowEvent;
+import edu.pdx.cs410J.ParserException;
+import edu.pdx.cs410J.grader.*;
+
+import javax.swing.*;
+import javax.swing.border.Border;
+import java.awt.*;
+import java.awt.event.*;
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintStream;
@@ -13,24 +14,6 @@ import java.net.URL;
 import java.util.ArrayList;
 import java.util.TimerTask;
 import java.util.prefs.Preferences;
-
-import javax.swing.AbstractAction;
-import javax.swing.Action;
-import javax.swing.BorderFactory;
-import javax.swing.ImageIcon;
-import javax.swing.JFileChooser;
-import javax.swing.JFrame;
-import javax.swing.JLabel;
-import javax.swing.JMenu;
-import javax.swing.JMenuBar;
-import javax.swing.JMenuItem;
-import javax.swing.JOptionPane;
-import javax.swing.JToolBar;
-import javax.swing.KeyStroke;
-import javax.swing.SwingUtilities;
-import javax.swing.border.Border;
-
-import edu.pdx.cs410J.ParserException;
 
 /**
  * This class is a main GUI for manipulate the grade book for CS410J.
